requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docx==0.2.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
pycodestyle==2.8.0
pydocstyle==6.1.1
PyQt3D==5.15.5
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.5
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.28.0
six==1.16.0
snowballstemmer==2.2.0
TBB==0.2
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
